refactor: Make layout::Resolution generic over platform
